Revert "Refactor: migrate communication services to notification module (#233)"
This reverts commit 7c1f40ff395dcfdf095dd46bf6fea8cba4829787.
